added commands for PTS and SRR, query for fetching followed sources and fixed map filters query

+++ backend/Main/Main/Queries/fetch_searched_sources/NewsSourceDto.cs
@@ -3,7 +3,7 @@ namespace Main.Queries.DTOs
     public class NewsSourceDto
     {
         public string SourceName { get; set; }
-        public int SRR { get; set; }
+        public double SRR { get; set; }
         public bool IsFollowing { get; set; }
     }
 }
